Добавил проверку при отнимании репутации (если бесконечно давать муты игроку то оно могло упасть ниже нуля

+++ src/main/java/ru/neroduckale/reputation/listener/PunishmentListener.java
@@ -21,6 +21,10 @@ public class PunishmentListener extends Events.Listener {
 
         int amount = configService.get("reputation-remove-amount");
 
+        if (userService.getUser(username).getReputation() < amount) {
+            return;
+        }
+
         userService.removeReputation(username, amount);
         PlayerUtil.sendMessage(username,
                 ((String) configService.get("get-punishment-message")).replace("{amount}", String.valueOf(amount))
